Refactor utils file
