+++ src/frontend/src/pages/FlowPage/components/PageComponent/utils/is-wrapped-with-class.tsx
@@ -1,0 +1,4 @@
+const isWrappedWithClass = (event: any, className: string | undefined) =>
+  event.target.closest(`.${className}`);
+
+export default isWrappedWithClass;
