+++ src/frontend/src/utils/reactflowUtils.ts
@@ -36,12 +36,8 @@ import {
   unselectAllNodesType,
   updateEdgesHandleIdsType,
 } from "../types/utils/reactflowUtils";
-import {
-  createRandomKey,
-  getFieldTitle,
-  getRandomDescription,
-  toTitleCase,
-} from "./utils";
+import { createRandomKey, getFieldTitle, toTitleCase } from "./utils";
+import { DESCRIPTIONS } from "../flow_constants";
 const uid = new ShortUniqueId({ length: 5 });
 
 export function checkChatInput(nodes: Node[]) {
@@ -1284,6 +1280,14 @@ export function downloadFlows() {
     // simulate a click on the link element to trigger the download
     link.click();
   });
+}
+
+export function getRandomElement<T>(array: T[]): T {
+  return array[Math.floor(Math.random() * array.length)];
+}
+
+export function getRandomDescription(): string {
+  return getRandomElement(DESCRIPTIONS);
 }
 
 export const createNewFlow = (
